Move order creation to setup and update tests
