+++ plugins/woocommerce/tests/performance/tests/gh-action-daily-ext-requests.js
@@ -21,6 +21,7 @@ import { ordersSearch } from '../requests/merchant/orders-search.js';
 import { addOrder } from '../requests/merchant/add-order.js';
 import { ordersAPI } from '../requests/api/orders.js';
 import { homeWCAdmin } from '../requests/merchant/home-wc-admin.js';
+import { addCustomerOrder } from '../setup/add-customer-order.js';
 
 const shopper_request_threshold = 'p(95)<10000';
 const merchant_request_threshold = 'p(95)<10000';
@@ -246,6 +247,10 @@ export const options = {
 		],
 	},
 };
+
+export function setup() {
+	addCustomerOrder();
+}
 
 export function shopperBrowseFlow() {
 	homePage();
