+++ plugins/woocommerce/tests/performance/tests/gh-action-pr-requests.js
@@ -23,6 +23,7 @@ import { addOrder } from '../requests/merchant/add-order.js';
 import { ordersAPI } from '../requests/api/orders.js';
 import { homeWCAdmin } from '../requests/merchant/home-wc-admin.js';
 import { setCartCheckoutShortcodes } from '../setup/cart-checkout-shortcode.js';
+import { addCustomerOrder } from '../setup/add-customer-order.js';
 
 const shopper_request_threshold = 'p(95)<10000';
 const merchant_request_threshold = 'p(95)<10000';
@@ -251,6 +252,7 @@ export const options = {
 
 export function setup() {
 	setCartCheckoutShortcodes();
+	addCustomerOrder();
 }
 
 export function shopperBrowseFlow() {
