+++ plugins/woocommerce/tests/performance/tests/hpos-baseline-load.js
@@ -25,6 +25,7 @@ import { myAccountMerchantLogin } from '../requests/merchant/my-account-merchant
 import { wpLogin } from '../requests/merchant/wp-login.js';
 import { ordersAPI } from '../requests/api/orders.js';
 import { admin_acc_login } from '../config.js';
+import { addCustomerOrder } from '../setup/add-customer-order.js';
 
 const shopper_request_threshold = 'p(95)<100000';
 const merchant_request_threshold = 'p(95)<100000';
@@ -256,6 +257,10 @@ export const options = {
 	},
 };
 
+export function setup() {
+	addCustomerOrder();
+}
+
 // Use myAccountMerchantLogin() instead of wpLogin() if having issues with login.
 export function merchantOrderFlows() {
 	if ( admin_acc_login === true ) {
@@ -277,7 +282,12 @@ export function merchantOtherFlows() {
 		wpLogin();
 	}
 	homeWCAdmin( { other: false, products: false, reviews: false } );
-	addProduct( { heartbeat: false, other: false, permalink: false, update: false } );
+	addProduct( {
+		heartbeat: false,
+		other: false,
+		permalink: false,
+		update: false,
+	} );
 	coupons();
 }
 export function shopperBrowsingFlows() {
